refactor: Lotto, WinningLotto 클래스 domain 폴더로 이동
