+++ __tests__/WinningLottoTest.js
@@ -1,4 +1,4 @@
-const WinningLotto = require("../src/WinningLotto");
+const WinningLotto = require("../src/domain/WinningLotto");
 
 describe("당첨 로또 클래스 테스트", () => {
   test("보너스 번호의 범위가 1~45 사이가 아닐 경우 예외가 발생한다.", () => {
